Move collision into Vector
