--- src/physics/Collision.js
@@ -1,32 +1,0 @@
-'use strict';
-
-var Vector = require('../graphics/Vector');
-var Point = require('../graphics/Point');
-
-module.exports = (vectorA, vectorB) => {
-
-  if (!Vector.Created(vectorA) || !Vector.Created(vectorB)) {
-    throw new Error('Invalid arguments passed to Collision, requires two Vectors');
-  }
-
-  var s1_x, s1_y, s2_x, s2_y;
-
-  s1_x = vectorA.to.x - vectorA.from.x;
-  s1_y = vectorA.to.y - vectorA.from.y;
-  s2_x = vectorB.to.x - vectorB.from.x;
-  s2_y = vectorB.to.y - vectorB.from.y;
-
-  var s = (-s1_y * (vectorA.from.x - vectorB.from.x) + s1_x * (vectorA.from.y - vectorB.from.y)) / (-s2_x * s1_y + s1_x * s2_y);
-  var t = (s2_x * (vectorA.from.y - vectorB.from.y) - s2_y * (vectorA.from.x - vectorB.from.x)) / (-s2_x * s1_y + s1_x * s2_y);
-
-  if (s >= 0 && s <= 1 && t >= 0 && t <= 1) {
-    // Collision detected
-    return Point({
-      x: vectorA.from.x + (t * s1_x),
-      y: vectorA.from.y + (t * s1_y)
-    });
-  }
-
-  // No collision
-  return false;
-};
